Move the load_tests function into project.py
The load_tests function has been moved from testloader.py into project.py and
the ChipToolsTest class has been updated so that it now expects a project path
on the class attribute "project" which it will attempt to load if the
environment is not aleady loaded by the ChipTools framework. This builds on the
earlier improvements to the ChipToolsTest class which added support for
external test runners. The user now only needs to specify a project path rather
than providing a reference to a fully loaded project when creating their tests.