Update ChipToolsTest
Update ChipToolsTest to allow for the case where user defined test classes
inherit from other user defined test classes and the project path for external
test runners is redefined. During setUpClass the ChipToolsTest now checks which
environment it is using ('chiptools' or 'external') and if 'external' it will
check to see if the project path is different, upon which it will reload the
new environment.